fixed the pipeline failures

+++ src/common-utils/main.sh
@@ -544,10 +544,14 @@ if [ "${INSTALL_ZSH}" = "true" ]; then
 
         # Add devcontainer .zshrc template
         if [ "$INSTALL_OH_MY_ZSH_CONFIG" = "true" ]; then
-            cp -f ${user_rc_file} ${copy_user_rc_file}
+            if [ -f "$user_rc_file" ]; then
+                cp -f ${user_rc_file} ${copy_user_rc_file}
+            fi
             echo -e "$(cat "${template_path}")\nDISABLE_AUTO_UPDATE=true\nDISABLE_UPDATE_PROMPT=true" > ${user_rc_file}
             sed -i -e 's/ZSH_THEME=.*/ZSH_THEME="devcontainers"/g' ${user_rc_file}
-            diff ${user_rc_file} ${copy_user_rc_file} | grep '^>' | sed 's/^>\ //' >> ${user_rc_file}
+            if [ -f "$user_rc_file" ]; then
+                diff ${user_rc_file} ${copy_user_rc_file} | grep '^>' | sed 's/^>\ //' >> ${user_rc_file}
+            fi
         fi
 
         # Copy to non-root user if one is specified
